+++ sharding-jdbc-core/src/main/java/com/dangdang/ddframe/rdb/sharding/router/PreparedSQLRouter.java
@@ -52,7 +52,7 @@ public class PreparedSQLRouter {
                 each.setNewConditionValue(parameters);
             }
         }
-        return engine.routeSQL(sqlParsedResult);
+        return engine.routeSQL(sqlParsedResult, parameters);
     }
 }