+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/merger/pipeline/reducer/IteratorResultSetTest.java
@@ -21,6 +21,7 @@ import com.dangdang.ddframe.rdb.sharding.merger.ResultSetFactory;
 import com.dangdang.ddframe.rdb.sharding.merger.fixture.MockResultSet;
 import com.dangdang.ddframe.rdb.sharding.parser.result.merger.Limit;
 import com.dangdang.ddframe.rdb.sharding.parser.result.merger.MergeContext;
+import com.google.common.base.Optional;
 import org.junit.Test;
 
 import java.sql.ResultSet;
@@ -46,7 +47,7 @@ public final class IteratorResultSetTest {
     @Test
     public void assertNextWithLimitForAllData() throws SQLException {
         MergeContext mergeContext = new MergeContext();
-        mergeContext.setLimit(new Limit(1, 10));
+        mergeContext.setLimit(new Limit(1, 10, Optional.<Integer>absent(), Optional.<Integer>absent()));
         ResultSet resultSet = ResultSetFactory.getResultSet(Arrays.<ResultSet>asList(new MockResultSet<>(1), new MockResultSet<>(2, 4), new MockResultSet<Integer>()), mergeContext);
         int count = 0;
         while (resultSet.next()) {
@@ -58,7 +59,7 @@ public final class IteratorResultSetTest {
     @Test
     public void assertNextWithLimitForPartData() throws SQLException {
         MergeContext mergeContext = new MergeContext();
-        mergeContext.setLimit(new Limit(1, 1));
+        mergeContext.setLimit(new Limit(1, 1, Optional.<Integer>absent(), Optional.<Integer>absent()));
         ResultSet resultSet = ResultSetFactory.getResultSet(Arrays.<ResultSet>asList(new MockResultSet<>(1), new MockResultSet<>(2, 4), new MockResultSet<Integer>()), mergeContext);
         int count = 0;
         while (resultSet.next()) {
@@ -70,7 +71,7 @@ public final class IteratorResultSetTest {
     @Test
     public void assertNextWithLimitForSingleResultSet() throws SQLException {
         MergeContext mergeContext = new MergeContext();
-        mergeContext.setLimit(new Limit(2, 3));
+        mergeContext.setLimit(new Limit(2, 3, Optional.<Integer>absent(), Optional.<Integer>absent()));
         ResultSet resultSet = ResultSetFactory.getResultSet(Collections.<ResultSet>singletonList(new MockResultSet<>(1, 2, 3, 4, 5, 6)), mergeContext);
         int count = 0;
         while (resultSet.next()) {
@@ -82,7 +83,7 @@ public final class IteratorResultSetTest {
     @Test
     public void assertNextWithLimitForSingleResultSetNotFull() throws SQLException {
         MergeContext mergeContext = new MergeContext();
-        mergeContext.setLimit(new Limit(2, 10));
+        mergeContext.setLimit(new Limit(2, 10, Optional.<Integer>absent(), Optional.<Integer>absent()));
         ResultSet resultSet = ResultSetFactory.getResultSet(Collections.<ResultSet>singletonList(new MockResultSet<>(1, 2, 3, 4, 5, 6)), mergeContext);
         int count = 0;
         while (resultSet.next()) {