+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/parser/jaxb/Limit.java
@@ -17,12 +17,14 @@
 
 package com.dangdang.ddframe.rdb.sharding.parser.jaxb;
 
+import com.google.common.base.Optional;
+import lombok.Getter;
+import lombok.Setter;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
-
-import lombok.Getter;
-import lombok.Setter;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 @Getter
 @Setter
@@ -34,4 +36,12 @@ public final class Limit {
     
     @XmlAttribute(name = "row-count")
     private Integer rowCount;
+    
+    @XmlJavaTypeAdapter(OptionalAdapter.class)
+    @XmlAttribute(name = "offset-index")
+    private Optional<Integer> offsetParameterIndex;
+    
+    @XmlJavaTypeAdapter(OptionalAdapter.class)
+    @XmlAttribute(name = "row-count-index")
+    private Optional<Integer> rowCountParameterIndex;
 }